requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
google-generativeai
python-dotenv==1.0.0
pydantic==1.10.9
httpx==0.24.1
google-genai
